gestion plano, falta corregir cliparea al max y min

+++ src/linea.cpp
@@ -47,7 +47,7 @@ string Linea::getCapa(){
     return capa;
 }
 
-void Linea::escalar(int escala){
+void Linea::escalar(double escala){
     if (escala>500||escala<0)
 	cerr<<"Error: escala incorrecta"<<endl;
     else{
